+++ src/features/logtalkTerminal.ts
@@ -398,6 +398,16 @@ export default class LogtalkTerminal {
     await LogtalkTerminal.waitForFile(marker);
   }
 
+  public static async getImplementations(doc: TextDocument, position: Position, predicate: string) {
+    LogtalkTerminal.createLogtalkTerm();
+    const dir0: string = LogtalkTerminal.ensureDir(doc.uri);
+    const dir = path.resolve(dir0).split(path.sep).join("/");
+    let goals = `vscode_reflection::find_implementations('${dir}', ${predicate}, '${doc.fileName}', ${position.line}).\r`;
+    LogtalkTerminal.sendString(goals);
+    const marker = path.join(dir0, ".implementations_done");
+    await LogtalkTerminal.waitForFile(marker);
+  }
+
   private static spawnScript4(dir: string, type: string[], path: string, args: string[]) {
     let pp = spawn(path, args, { cwd: dir })
       .on("stdout", out => {