+++ src/utils/utils.ts
@@ -76,6 +76,24 @@ export class Utils {
     return desc;
   }
 
+  public static getPredicateIndicatorUnderCursor(
+    doc: TextDocument,
+    position: Position
+  ): string {
+    let wordRange: Range = doc.getWordRangeAtPosition(
+      position,
+      /(\w+)[/](\d+)/
+    );
+    if (!wordRange) {
+      return null;
+    }
+    let doctext = doc.getText(wordRange);
+    let match = doctext.match(/(\w+)[/](\d+)/);
+    let name: string = match[1];
+    let arity: number = parseInt(match[2]);
+    return name + "/" + arity;
+  }
+
   public static getPredicateUnderCursor(
     doc: TextDocument,
     position: Position