+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add "Go to Declaration" and "Go to Definition" support
 * Add "Go to Type Definition" support (with objects, protocols, and categories interpreted as types)
+* Add "Go to References" and "Go to Implementations" support
 * Fix JavaScript error when parsing terminal output
 
 ## [0.19.0]